moving to doubles, refs #369
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@10125 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/tools/traci/inductionloop.py
@@ -33,11 +33,11 @@ def readVehicleData(result):
 
 RETURN_VALUE_FUNC = {tc.ID_LIST:                        traci.Storage.readStringList,
                      tc.LAST_STEP_VEHICLE_NUMBER:       traci.Storage.readInt,
-                     tc.LAST_STEP_MEAN_SPEED:           traci.Storage.readFloat,
+                     tc.LAST_STEP_MEAN_SPEED:           traci.Storage.readDouble,
                      tc.LAST_STEP_VEHICLE_ID_LIST:      traci.Storage.readStringList,
-                     tc.LAST_STEP_OCCUPANCY:            traci.Storage.readFloat,
-                     tc.LAST_STEP_LENGTH:               traci.Storage.readFloat,
-                     tc.LAST_STEP_TIME_SINCE_DETECTION: traci.Storage.readFloat,
+                     tc.LAST_STEP_OCCUPANCY:            traci.Storage.readDouble,
+                     tc.LAST_STEP_LENGTH:               traci.Storage.readDouble,
+                     tc.LAST_STEP_TIME_SINCE_DETECTION: traci.Storage.readDouble,
                      tc.LAST_STEP_VEHICLE_DATA:         readVehicleData}
 subscriptionResults = {}
